Help

+++ HotelManagerBackend/src/main/java/HotelManager/BusinessLayer/ErrorHandling/AccountNotFoundException.java
@@ -2,7 +2,7 @@ package HotelManager.BusinessLayer.ErrorHandling;
 
 public class AccountNotFoundException extends RuntimeException {
 
-    public AccountNotFoundException(Long id) {
+    public AccountNotFoundException(Integer id) {
         super("Could not find Account with ID: " + id);
     }
 }
